Finish renaming API
