+++ python/ray/util/sgd/torch/__init__.py
@@ -2,17 +2,17 @@ import logging
 logger = logging.getLogger(__name__)
 
 TorchTrainer = None
-PyTorchTrainable = None
+TorchTrainable = None
 TrainingOperator = None
 
 try:
     import torch  # noqa: F401
 
-    from ray.util.sgd.pytorch.pytorch_trainer import (TorchTrainer,
-                                                      PyTorchTrainable)
+    from ray.util.sgd.torch.torch_trainer import (TorchTrainer,
+                                                      TorchTrainable)
 
-    from ray.util.sgd.pytorch.training_operator import TrainingOperator
+    from ray.util.sgd.torch.training_operator import TrainingOperator
 
-    __all__ = ["TorchTrainer", "PyTorchTrainable", "TrainingOperator"]
+    __all__ = ["TorchTrainer", "TorchTrainable", "TrainingOperator"]
 except ImportError:
     logger.warning("PyTorch not found. TorchTrainer will not be available")
