+++ python/ray/util/sgd/torch/examples/cifar_pytorch_example.py
@@ -8,8 +8,8 @@ import torchvision
 import torchvision.transforms as transforms
 
 import ray
-from ray.util.sgd.pytorch import (TorchTrainer, PyTorchTrainable)
-from ray.util.sgd.pytorch.resnet import ResNet18
+from ray.util.sgd.torch import (TorchTrainer, TorchTrainable)
+from ray.util.sgd.torch.resnet import ResNet18
 
 
 def initialization_hook():
@@ -107,7 +107,7 @@ def tune_example(num_replicas=1, use_gpu=False, test_mode=False):
     }
 
     analysis = tune.run(
-        PyTorchTrainable,
+        TorchTrainable,
         num_samples=2,
         config=config,
         stop={"training_iteration": 2},
