+++ python/ray/util/sgd/torch/torch_trainer.py
@@ -11,11 +11,11 @@ import ray
 
 from ray.tune import Trainable
 from ray.tune.trial import Resources
-from ray.util.sgd.pytorch.distributed_pytorch_runner import (
+from ray.util.sgd.torch.distributed_torch_runner import (
     DistributedTorchRunner)
 from ray.util.sgd import utils
-from ray.util.sgd.pytorch.pytorch_runner import TorchRunner
-from ray.util.sgd.pytorch.constants import VALID_SCHEDULER_STEP
+from ray.util.sgd.torch.torch_runner import TorchRunner
+from ray.util.sgd.torch.constants import VALID_SCHEDULER_STEP
 
 logger = logging.getLogger(__name__)
 RESIZE_COOLDOWN_S = 10
@@ -487,7 +487,7 @@ class TorchTrainer:
         return False
 
 
-class PyTorchTrainable(Trainable):
+class TorchTrainable(Trainable):
     @classmethod
     def default_resource_request(cls, config):
         return Resources(
