+++ python/ray/util/sgd/torch/training_operator.py
@@ -2,7 +2,7 @@ import collections
 import torch
 
 from ray.util.sgd.utils import TimerStat, AverageMeter
-from ray.util.sgd.pytorch.constants import (
+from ray.util.sgd.torch.constants import (
     SCHEDULER_STEP_EPOCH, SCHEDULER_STEP_BATCH, SCHEDULER_STEP, BATCH_COUNT)
 
 amp = None
@@ -11,7 +11,7 @@ try:
     from apex import amp
 except ImportError:
     # Apex library is not installed, so we cannot enable mixed precision.
-    # We don't log here because logging happens in the pytorch_runner,
+    # We don't log here because logging happens in the torch_runner,
     # where amp is initialized.
     pass
